requirements.txt
aiohttp==1.0.5
asn1crypto==0.23.0
async-timeout==2.0.0
attrs==17.3.0
Automat==0.6.0
beautifulsoup4==4.6.0
cffi==1.11.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.1.3
discord.py==0.16.12
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
multidict==3.3.2
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycparser==2.18
pyOpenSSL==17.3.0
service-identity==17.0.0
six==1.11.0
websockets==3.4
yarl==0.14.0
zope.interface==4.4.3
